--- front-end/src/components/legacy/Topbar/Topbar.module.scss
@@ -1,62 +1,0 @@
-@use '@styles/_fonts';
-
-.topbar {
-    @include fonts.title-medium;
-    background: var(--background);
-    display: flex;
-    justify-content: space-between;
-    padding: 1rem 0rem 1rem 0rem;
-    color: var(--text-colour);
-    position: sticky;
-    z-index: 1;
-    width: 100%;
-    height: 3rem;
-
-    > * {
-        margin: 0 1rem 0 1rem;
-    }
-
-    .title {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        cursor: pointer;
-
-        img {
-            width: 150px !important;
-        }
-    }
-
-    .buttons {
-        display: flex;
-
-        > * {
-            padding: 8px 5px;
-            font-family: "HelveticaNeueW01-77BdCnObl";
-            background: none !important;
-            width: fit-content; 
-            font-size: 2.5rem;
-            margin: 0px 8px !important;
-            align-items: center;
-            display: flex;
-
-            &:hover {
-                filter: brightness(0.8);
-            }
-
-            &:active {
-                box-shadow: 0px 0px 3px #000000;
-                transition: box-shadow 1s;
-                transition-delay: -0.5s;
-            }
-
-            .icon {
-                width: 1em;
-                height: 1em;
-                display: flex;
-            }
-        }
-    }
-}
-
-
