--- front-end/src/components/legacy/modal/FileStorage/FileStorage.css
@@ -1,126 +1,0 @@
-.file-Storage-Container {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-}
-
-.file-browser {
-    font-size: 1.2rem;
-    line-height: 1.8rem;
-    padding: 0;
-    margin: 0;
-    font-family: 'Lato', sans-serif;
-    font-weight: 200;
-    word-break: break-all;
-}
-
-.file.selected {
-    background-color: rgba(0, 123, 255, 0.25);
-}
-
-a {
-    color: #425363;
-    text-decoration: none;
-}
-
-a:hover {
-    color: #192026;
-    text-decoration: none;
-}
-
-ul li,
-ol li {
-    margin: 0.5rem 0;
-}
-
-code {
-    max-height: 25vh;
-    white-space: pre;
-    display: block;
-    background: #C6EAFB;
-    padding: 1rem;
-    overflow-y: auto;
-    color: #000;
-}
-
-div.content-container {
-    max-width: 50rem;
-    margin: 0 auto;
-}
-
-div.header {
-    padding: 2rem;
-    background: #00ADEF;
-    color: #fff;
-}
-
-div.header a {
-    color: #fff;
-}
-
-div.header a.title {
-    font-weight: 400;
-    font-size: 1.6rem;
-}
-
-div.content {
-    background: #fff;
-    padding: 4rem 1rem;
-}
-
-a.source-link {
-    display: block;
-    color: #000;
-    text-align: right;
-    text-transform: uppercase;
-    font-size: 0.8rem;
-    line-height: 1em;
-    margin: 0.5rem 0;
-}
-
-div.footer {
-    padding: 3rem 1rem;
-    color: #fff;
-}
-
-.rendered-react-keyed-file-browser .rendered-file-browser .files {
-    max-height: 300px;
-    overflow-y: auto;
-    display: block;
-}
-
-.fileButtons {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    margin-top: auto;
-}
-
-.pageTwoNextButton {
-    padding: 13px 29px;
-    font-size: 1rem;
-    font-weight: bold;
-    border-radius: 5px;
-    border: none;
-    outline: none;
-    background: crimson;
-    color: white;
-    cursor: pointer;
-}
-
-.pageTwoBackButton {
-    padding: 13px 29px;
-    font-size: 1rem;
-    font-weight: bold;
-    border-radius: 5px;
-    border: none;
-    outline: none;
-    background: crimson;
-    color: white;
-    cursor: pointer;
-}
-
-.name {
-    max-width: 250px
-}
