--- front-end/src/components/legacy/modal/create/AnalyzersAndSeries/AnalyzersAndSeries.css
@@ -1,91 +1,0 @@
-.columnHeaders {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-}
-
-/* TODO: For the love of god, don't use this name.
-   TODO: This will be replaced when our header selection on the last page is replaced, but these need to be fixed baad */
-.columnHeaders2 {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.analyzersAndSeriesContainer {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: left;
-  flex: 1;
-}
-
-
-
-.analyzerContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: left;
-  overflow: auto;
-  background-color: #f1f1f1;
-  margin-bottom: 2em;
-}
-
-.analyzerBox {
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  margin: 0.5em 0;
-}
-
-.boldText {
-  font-weight: bold;
-  margin: 0 0 0.25em 0.5em;
-}
-
-.buttonFlexBox {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  margin-top: auto;
-}
-
-
-
-details {
-  display: block;
-}
-
-details>summary {
-  list-style: none;
-  display: inline-block;
-}
-
-summary::after {
-  content: ' ▶';
-}
-
-details[open] summary:after {
-  content: " ▼";
-}
-
-.info {
-  display: flex;
-  margin-left: auto;
-}
-
-/* Add a background color to the button if it is clicked on (add the .active class with JS), and when you move the mouse over it (hover) */
-.active,
-.collapsible:hover {
-  background-color: #ccc;
-}
-
-/* Style the collapsible content. Note: hidden by default */
-.content {
-  padding: 0 18px;
-  display: none;
-  overflow: hidden;
-  background-color: #f1f1f1;
-}
-
