--- front-end/src/components/legacy/modal/create/PresetSelectionPage/PresetSelectionPage.css
@@ -1,34 +1,0 @@
-.presets-container {
-    display: grid;
-    grid-template-columns: auto auto auto;
-    gap: 8px;
-}
-
-.presets-instructions {
-    grid-column: 1 / -1;
-}
-
-.PageButton {
-    grid-column: 1 / -1;
-}
-
-.presets-preset {
-    padding-left: 4px;
-    padding-right: 4px;
-    border: 1px solid black;
-    border-radius: 4%;
-}
-
-.presets-preset--selected {
-    border: 3px solid black;
-}
-
-.presets-preset-title {
-    font-size: 24px;
-    margin: 3px;
-}
-
-.presets-preset-description {
-    font-size: 14px;
-    margin: 2px
-}
