--- front-end/src/components/legacy/modal/create/VideoSelect/VideoSelect.css
@@ -1,76 +1,0 @@
-.videoSelectContainer {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: left;
-    flex: 1;
-  }
-
-h3 {
-  font-size: x-large;
-}
-
-.fileButtons {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    margin-top: auto;
-}
-
-.videoSelect {
-    height: 100%;
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    flex: 1;
-    margin-bottom: 30px;
-    max-height: 340px;
-  }
-
-.videoContainer {
-    flex: 2.7;
-    display: flex;
-    flex-direction: column;
-    border: 2px solid black;
-    margin-right: 10px;
-    overflow-y: auto;
-}
-
-.videoLabel {
-    cursor: pointer;
-    transition: background-color 0.3s;
-    padding: 10px;
-  }
-
-.videoLabel:hover {
-    background-color: #e0e0e0;
-  }
-
-.videoLabel.selected {
-    background-color: #3498db;
-  }
-
-.videoLabel input[type="radio"] {
-    display: none;
-    }
-
-.folderContainer {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    border: 2px solid black;
-    padding-top: 10px;
-    align-items: center;
-    overflow-y: auto;
-}
-
-.folderLabel {
-    margin-bottom: 15px;
-}
-
-.folderContainerLabel {
-    font-weight: 600;
-    text-align: center;
-    margin-bottom: 10px;
-}
-
