--- front-end/src/components/legacy/modal/upload/UploadModal.css
@@ -1,92 +1,0 @@
-
-#form-file-upload {
-  height: 16rem;
-  width: 28rem;
-  max-width: 100%;
-  text-align: center;
-  position: relative;
-}
-
-#input-file-upload {
-  display: none;
-}
-
-#label-file-upload {
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-width: 2px;
-  border-radius: 1rem;
-  border-style: dashed;
-  border-color: #cbd5e1;
-  background-color: #f8fafc;
-}
-
-#label-file-upload.drag-active {
-  background-color: #e5e7eb;
-}
-
-.upload-button {
-  cursor: pointer;
-  padding: 0.25rem;
-  font-size: 1rem;
-  border: none;
-  /* font-family: 'Oswald', sans-serif; */
-  background-color: transparent;
-}
-
-.upload-button:hover {
-  text-decoration-line: underline;
-}
-
-#drag-file-element {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  border-radius: 1rem;
-  top: 0px;
-  right: 0px;
-  bottom: 0px;
-  left: 0px;
-}
-
-
-.submitbutton:active {
-  outline: none;
-  background: linear-gradient(0deg, rgba(0, 0, 0, 0.2), rgba(255, 255, 255, 0.1)),crimson;
-}
-
-.loading {
-  position: absolute;
-  left: 0; top: 5rem; right: 0; bottom: 0;
-  width: 28rem;
-}
-
-
-.uploadContainer{
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.uploadSubmit {
-  margin: 1em;
-  padding: 13px 29px;
-  font-size: 1rem;
-  font-weight: bold;
-  border-radius: 5px;
-  border: none;
-  outline: none;
-  background: crimson;
-  color: white;
-  cursor: pointer;
-}
-
-.fileButton{
-  color: white;
-  background-color: red;
-  margin: 0;
-  padding: 7px 7px;
-}
