+++ front-end/src/index.tsx
@@ -1,6 +1,6 @@
 import ReactDOM from 'react-dom/client';
 import '@styles/index.scss';
-import App from '@components/legacy/App/App';
+import App from '@pages/App/App';
 import { ThemeProvider } from '@contexts/ThemeContext';
 import { ModalProvider } from '@contexts/ModalContext';
 import { RouterComponent } from '@lib/navigationUtils';
