+++ front-end/src/pages/App/App.module.scss
@@ -1,0 +1,9 @@
+.App {
+  display: flex;
+  height: 100vh;
+
+  .Body {
+    width: 100%;
+    height: 100%;
+  }
+}
